Fixes for next.js
This throws an error in next.js because instanceOf(Element) will not work for SSR.
The usage of dynamic import isn't exactly a solution but a workaround to a problem that only involves typechecking. 
Since this is for a react library, PropTypes.element should be good to check for the typing.